+++ apps/server/src/app.module.ts
@@ -4,7 +4,7 @@ import { PrismaService } from './prisma/prisma.service';
 import { SharedModule } from './shared/shared.module';
 import { ProgramsModule } from './programs/programs.module';
 import { SeedGroupsModule } from './seed-groups/seed-groups.module';
-import { MembersModule } from './members/members.module';
+import { PersonModule } from './members/person.module';
 
 @Module({
   imports: [
@@ -12,7 +12,7 @@ import { MembersModule } from './members/members.module';
     SharedModule,
     ProgramsModule,
     SeedGroupsModule,
-    MembersModule,
+    PersonModule,
   ],
   providers: [PrismaService],
   exports: [PrismaService],
